add metrics weight selectors

+++ src/MultiCloudAISelectionPlatform/MultiCloudAISelectionPlatform.Common/Models/MetrikWeights.cs
@@ -2,8 +2,8 @@
 {
     public class MetrikWeights
     {
-        public int Accuracy { get; set; } = 1;
-        public int Costs { get; set; } = 1;
-        public int ResponseTime { get; set; } = 1;
+        public double Accuracy { get; set; } = 1.0;
+        public double Costs { get; set; } = 1.0;
+        public double ResponseTime { get; set; } = 1.0;
     }
 }
